refactor(roxagent): extract discovery package from vsock client
Prompt: PR-splitting follow-up on stackrox/stackrox#21435 ("Do you see
any low-hanging fruits how we can split this PR?" / "I like the
discovery extraction").

Moves OS/activation/DNF-metadata discovery out of vsock/discovery.go
(private to the push-mode client) into its own discovery package, so
the new pull-mode roxagent-serve command (landing separately) can
reuse it instead of duplicating it. vsock/client.go is updated to call
discovery.DiscoverVMData; behavior for the existing push-mode path is
unchanged.

discovery.go now defines its own package-level `log`, previously
shared implicitly via vsock/client.go within the same package.

Partially generated with AI assistance.
