+++ compliance/virtualmachines/roxagent/discovery/discovery_test.go
@@ -1,4 +1,4 @@
-package vsock
+package discovery
 
 import (
 	"os"
